requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
langgraph>=0.0.40
firecrawl-py>=0.0.20
langchain-openai>=0.1.7
python-dotenv>=1.0.1
langchain-ollama